requirements.txt
scipy
numpy
matplotlib
pandas
statsmodels
sklearn
keras
tensorflow==1.5.0
deepspeech
jupyter
pillow
python-resize-image
